Move generator code in common class
