+++ reactive-grpc-common/pom.xml
@@ -25,6 +25,10 @@
             <artifactId>grpc-contrib</artifactId>
         </dependency>
         <dependency>
+            <groupId>com.salesforce.servicelibs</groupId>
+            <artifactId>jprotoc</artifactId>
+        </dependency>
+        <dependency>
             <groupId>org.reactivestreams</groupId>
             <artifactId>reactive-streams</artifactId>
             <version>1.0.1</version>
