+++ reactor/reactor-grpc-stub/pom.xml
@@ -21,10 +21,6 @@
 
     <dependencies>
         <dependency>
-            <groupId>com.salesforce.servicelibs</groupId>
-            <artifactId>grpc-contrib</artifactId>
-        </dependency>
-        <dependency>
             <groupId>${project.groupId}</groupId>
             <artifactId>reactive-grpc-common</artifactId>
             <version>${project.version}</version>
